refactor : bookItem, bookItemWrapper, bookItemSkeleton, etc. >> >> Related to : #147
