+++ src/app/theme/[slug]/page.tsx
@@ -1,8 +1,8 @@
 import ThemeHeader from '@/components/theme/ThemeHeader';
 import { recommendedData } from '@/apis/theme/recommendedData';
 import styles from '@/styles/theme/artist/artist.module.css';
-import BookItem from '@/components/common/BookItem';
 import { RecommendedDataType } from '@/types/bookType';
+import BookItem from '@/components/common/bookItem/BookItem';
 
 interface PageParams {
 	params: {
@@ -17,7 +17,7 @@ export default function page({ params }: PageParams) {
 
 	return (
 		<div>
-			<ThemeHeader title={data.tag} color={data.color} image={data.image} />
+			<ThemeHeader title={data.tag} />
 			<div className={styles.container}>
 				<div className={styles.wrapper}>
 					{data.book.map((book: any) => (
