+++ src/app/used/page.tsx
@@ -1,10 +1,10 @@
 'use client';
 
 import BestSeller from '@/components/common/BestSeller';
-import BookItemWrapper from '@/components/common/BookItemWrapper';
 import CategoryBox from '@/components/common/CategoryBox';
 import PageHeader from '@/components/common/PageHeader';
 import Pagination from '@/components/common/Pagination';
+import BookItemWrapper from '@/components/common/bookItem/BookItemWrapper';
 import RecentlyViewedBooks from '@/components/layout/RecentlyViewedBooks';
 import styles from '@/styles/used/used.module.css';
 import { UsedBookType } from '@/types/bookType';
