+++ src/components/common/bookItem/BookItemSkeleton.tsx
@@ -1,0 +1,3 @@
+export default function BookItemSkeleton() {
+	return <div></div>;
+}
