+++ src/components/search/SearchBookItemWrapper.tsx
@@ -1,7 +1,7 @@
 import { NewBookType } from '@/types/bookType';
 import React from 'react';
-import BookItem from '../common/BookItem';
 import styles from '@/styles/search/searchBookItemWrapper.module.css';
+import BookItem from '../common/bookItem/BookItem';
 
 interface SearchBookItemWrapperProp {
 	data: NewBookType[];
