+++ src/styles/cart/cartView/cartHeader/cartHeader.module.css
@@ -3,7 +3,6 @@
 	height: 50px;
 	display: flex;
 	align-items: center;
-	font-size: 14px;
 	text-align: center;
 
 	border-top: 1px solid #a9a9a9;
