+++ src/styles/cart/cartView/orderButton/orderButton.module.css
@@ -4,7 +4,7 @@
 	justify-content: center;
 	align-items: center;
 
-	margin: 30px 0 30px 0;
+	margin: 50px 0 50px 0;
 }
 .orderButtonWrap > a > button {
 	width: 125px;
