+++ src/styles/cart/cartView/orderHistory/orderHistory.module.css
@@ -8,8 +8,8 @@
 }
 .orderHistoryBody {
 	width: 30%;
-	margin: 30px 0 30px 0;
-	gap: 20px;
+	margin: 40px 0 40px 0;
+	gap: 25px;
 
 	display: flex;
 	flex-direction: column;
