--- src/styles/common/bookItem.module.css
@@ -1,94 +1,0 @@
-/* 도서 컴포넌트 */
-.bookItem {
-	width: 200px;
-	height: 380px;
-
-	display: flex;
-	flex-direction: column;
-	position: relative;
-
-	cursor: pointer;
-}
-
-.bookImage {
-	width: inherit;
-	height: 275px;
-	border-radius: 10px;
-
-	position: relative;
-	overflow: hidden;
-	transition: transform 0.3s ease-in-out;
-}
-
-.bookImage::after {
-	/* 이미지 위 오버레이 효과 */
-	content: '';
-	/* 레이어 자체 클릭 이벤트를 처리하지 않게하는 효과 */
-	pointer-events: none;
-
-	position: absolute;
-	top: 0;
-	left: 0;
-	right: 0;
-	bottom: 0;
-	background-color: rgba(0, 0, 0, 0.02);
-	border-radius: 10px;
-}
-.bookImage:hover {
-	transform: scale(1.05);
-}
-.bookText {
-	width: inherit;
-	height: 85px;
-
-	margin-top: 20px;
-}
-
-.bookText h1 {
-	height: 34px;
-	font-size: 16px;
-
-	overflow: hidden;
-	text-overflow: ellipsis;
-	display: -webkit-box;
-	-webkit-line-clamp: 2;
-	-webkit-box-orient: vertical;
-}
-.bookText a {
-	color: #000000;
-}
-.bookText a:hover {
-	text-decoration: underline;
-}
-.bookText p {
-	margin-top: 5px;
-	height: 15px;
-	font-size: 12px;
-
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-}
-.bookText p:nth-of-type(1) {
-	margin-top: 16px;
-	color: #aaaaaa;
-}
-.addDeleteBook {
-	position: absolute;
-	right: 8px;
-	bottom: 120px;
-	cursor: pointer;
-}
-
-/* 랭크 숫자 스타일*/
-.rank {
-	position: absolute;
-	top: 140px;
-	left: -30px;
-	font-size: 150px;
-	padding: 5px;
-	z-index: 1;
-	text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.5);
-
-	color: #ffffff;
-}
