+++ src/styles/common/bookItem/bookItemWrapper.module.css
@@ -1,0 +1,8 @@
+.bookItemWrapper {
+	width: inherit;
+	column-gap: 20px;
+	row-gap: 50px;
+
+	display: grid;
+	grid-template-columns: repeat(6, 1fr);
+}
