+++ src/styles/theme/artist/artist.module.css
@@ -5,7 +5,7 @@
 .wrapper {
 	width: 1300px;
 	gap: 75px;
-	margin: 50px 0 50px 0;
+	margin: 100px 0 100px 0;
 
 	display: grid;
 	grid-template-columns: repeat(5, 1fr);
