+++ src/styles/theme/theme.module.css
@@ -8,7 +8,7 @@
 .wrapper {
 	width: 1300px;
 	gap: 50px;
-	margin: 50px 0 50px 0;
+	margin: 100px 0 100px 0;
 
 	display: grid;
 	grid-template-columns: repeat(3, 1fr);
